Taproots API

Taproots — Module

Taproots is a library for traversing nested data structures.

Module Index

Taproots.Taproot

Taproots.adjacencymatrix

Taproots.bloom

Taproots.bottomup

Taproots.branches

Taproots.branchmap

Taproots.branchmap!

Taproots.branchprune

Taproots.branchprune!

Taproots.children

Taproots.data

Taproots.eachfield

Taproots.findtrace

Taproots.findtraces

Taproots.followindexes

Taproots.followtrace

Taproots.isbranch

Taproots.ischild

Taproots.isleaf

Taproots.isparent

Taproots.leafmap

Taproots.leafmap!

Taproots.leafprune

Taproots.leafprune!

Taproots.leaves

Taproots.postorder

Taproots.preorder

Taproots.prune

Taproots.prune!

Taproots.setchildren!

Taproots.setdata!

Taproots.tapin

Taproots.tapmap

Taproots.tapmap!

Taproots.tapmapif

Taproots.tapmapif!

Taproots.tapout

Taproots.topdown

Taproots.tracepairs

Taproots.traces

Taproots.@sprout

API

Taproots.Taproot — Type

A Taproot is a single node that that can contain other Taproots. It can itself be contained in a parent Taproot. It is not the main point of this package, but provides a convenience to persist structure when doing maps or other transformations. It is also useful if you simply wish to build a data structure which obeys everything in Taproots.jl.

You can construct a Taproot either by:

Taproot(data, children) where data is anything that you'd like to store in this node, and children is an iterable list of children

Taproot(node) where node is some value that has Taproots.children and Taproots.data defined on it. This will recurvisely sink your object into a Taproot

Taproots.adjacencymatrix — Method

adjacencymatrix(taproot)

Returns an adjacency matrix for this taproot and all its children. If two nodes are equivalent they will only appear once here.

Taproots.bloom — Method

bloom(io::IO, taproot)

Pretty print a taproot. Requires you to have used @sprout YourType for it to work.

Taproots.bottomup — Method

bottomup(x)

This creates an eager iterator as a bottomup (reverse level-order) breadth-first search of your custom Taproot. The bottom children are iterated first, and then the next layer up and so on. Leaves are not technically guaranteed to be iterated on first, before parents. While the other iterators are lazy, this one is eager, and will actually return a vector.

Usage

for x in bottomup(x) print(x) end

bottomup(x) <: Vector collect(bottomup(x)) <: Vector

Taproots.branches — Method

branches(x)

This creates a lazy iterator for the branches (those nodes which have children) of your custom Taproot.

Usage

for x in branches(x) print(x) end

collect(branches(x)) <: Vector

Taproots.branchmap! — Method

branchmap!(f::Function, taproot)

Modify the branches of a taproot in place without destroying links to children.

Taproots.branchmap — Method

branchmap(f::Function, taproot)

Deepcopy a taproot and then modify its branches in place without destroying links to children.

Taproots.branchprune! — Method

branchprune!(f::Function, taproot)

This removes any leaves of the taproot who do not satisfy the criteria given by f in place. f evaluating to true will keep that leaf. This is similar to filter. However, you cannot prune the root of a taproot.

Taproots.branchprune — Method

prune(f::Function, taproot)

This deepcopies the taproot and then removes any leaves who do not satisfy the criteria given by f in place. f evaluating to true will keep that leaf. This is similar to filter. However, you cannot prune the root of a taproot.

Taproots.children — Method

children(x)

Gets the children nodes in a Taproot DAG (if defined) of x. This function is meant to be overridden to access Taproot functionality on your own types.

By default, children(x) will return an empty array, meaning Taproot considers x to be a graph with a singular node (unless you overload this function.)

Overriding this function

Simply define Taproots.children(x::MyType) = ... Where the ... returns a vector of the things you want to traverse.

Taproots.data — Method

data(x)

Gets the data of a single node in a Taproot DAG (if defined) of x. This function is meant to be overridden to access Taproot functionality on your own types.

By default, data(x) will return x, meaning Taproots considers x to be perfectly valid data.

Overriding this function

Simply define Taproots.data(x::MyType) = ... Where the ... returns whatever data is not already included in the Taproots.children function.

Taproots.eachfield — Method

eachfield(x)

Returns a tuple of the values in each field in a struct. Nice for if you have a known number of children.

Taproots.findtrace — Method

findtrace(f, parent)

This is a slower (~linear) algorithm to find the first value where f returns true. It returns a vector which can be used to index into a nested data struct. You can index into your nested data structure using followtrace.

Taproots.findtraces — Method

findtraces(f, parent)

This is a slower (~linear) algorithm to find all values where f returns true. It returns a vector which can be used to index into a nested data struct. You can index into your nested data structure using followtrace.

Taproots.followindexes — Function

followindexes(indexable_struct, trace, default_value)

This is just a convenience function to index into a nest object (potentially not a taproot in any way) by some iterable value of keys.

Example

dict = Dict(1 => Dict(:a => Dict(1 => Dict(:a => "Finally here"))))

followindexes(dict, (1,:a,1,:a)) == "Finally here"

Taproots.followtrace — Method

followtrace(parent, trace)

This is a faster algorithm to get the node which matches a trace. A trace is simply an iterable set of indices.

Taproots.isbranch — Method

isbranch(node)::Bool

Tells you whether this node is a branch node (i.e. it has children).

Taproots.ischild — Method

ischild(potential_child, parent)::Bool

Tells you whether potential_child is a child of parent. This function acts recursively and is not the same as doing potential_child ∈ children(parent). Time complexity of this is ~linear in the number of edges in your DAG underneath the parent, so use with care for large DAGs.

Taproots.isleaf — Method

isleaf(node)::Bool

Tells you whether this node is a leaf node (i.e. it has no children).

Taproots.isparent — Method

isparent(potential_parent, child)::Bool

Tells you whether potential_parent is a parent of child. This function acts recursively. Time complexity of this is ~linear in the number of edges in your DAG underneath the potential_parent, so use with care for large DAGs.

Taproots.leafmap! — Method

leafmap!(f::Function, taproot)

Modify the leaves of a taproot in place.

Taproots.leafmap — Method

leafmap(f::Function, taproot)

Deepcopy a taproot and then modify its leaves in place.

Taproots.leafprune! — Method

leafprune!(f::Function, taproot)

This removes any leaves of the taproot who do not satisfy the criteria given by f in place. f evaluating to true will keep that leaf. This is similar to filter. However, you cannot prune the root of a taproot.

Taproots.leafprune — Method

prune(f::Function, taproot)

This deepcopies the taproot and then removes any leaves who do not satisfy the criteria given by f in place. f evaluating to true will keep that leaf. This is similar to filter. However, you cannot prune the root of a taproot.

Taproots.leaves — Method

leaves(x)

This creates a lazy iterator for the leaves (those nodes which have no children) of your custom Taproot.

Usage

for x in leaves(x) print(x) end

collect(leaves(x)) <: Vector

Taproots.postorder — Method

postorder(x)

This creates a lazy iterator as a postorder depth-first search of your custom Taproot. This is topological order and most of the time, you're looking for this. In this iterator, children are always iterated on before their parents. Parents will immediately follow their children (so some (possibly irrelevant) children may not be iterated before the parent above them).

Usage

for x in postorder(x) print(x) end

collect(postorder(x)) <: Vector

Taproots.preorder — Method

preorder(x)

This creates a lazy iterator as a preorder depth-first search of your custom Taproot. In this iterator, parents are always iterated on before their children.

Usage

for x in preorder(x) print(x) end

collect(preorder(x)) <: Vector

Taproots.prune! — Method

prune!(f::Function, taproot)

This removes any children who do not satisfy the criteria given by f in place. f evaluating to true will keep that child. This is similar to filter. However, you cannot prune the root of a taproot.

Taproots.prune — Method

prune(f::Function, taproot)

This deepcopies the taproot and then removes any children who do not satisfy the criteria given by f in place. f evaluating to true will keep that child. This is similar to filter. However, you cannot prune the root of a taproot.

Taproots.setchildren! — Method

setchildren!(node, children::Vector)

Sets the children of your node. Useful if you want to be able to prune! your taproot! Should return the entire node once completed.

Taproots.setdata! — Method

setdata!(node, data)

Sets the data of your node. Useful if you want to be able to tapmap! your taproot without messing up your children. It should be the case that the following does nothing. setdata!(origvalue, data(origvalue)) It MUST return the entire node once done (and is allowed to simply return data without modifying the node in place).

Example

Taproots.setdata!(node::MyDataStructure, data) = (node.data = data; node)

Taproots.tapin — Method

tapin(node)

Another way to access the recursive Taproot constructor, using a custom sink to collect data and children. Requires Taproots.data and Taproots.children to work.

Taproots.tapmap! — Method

tapmap!(f::Function, taproot)

Modify all the data of every node in a taproot in place. f only acts on the data in each node. This can handle mutable and immutable trees.

Taproots.tapmap — Method

tapmap(f::Function, taproot)

Deepcopy a taproot and then modify its nodes in place. f only acts on the data in each node. This can handle mutable and immutable trees.

Taproots.tapmapif! — Method

tapmapif!(condition::Function, f::Function, taproot)

Modifies nodes in place if the entire node satisfies condition. f only acts on the data in each node. This can handle mutable and immutable trees.

Taproots.tapmapif — Method

tapmapif(condition::Function, f::Function, taproot)

Deepcopy a taproot and then modify its nodes in place if they satisfy condition. f only acts on the data in each node. This can handle mutable and immutable trees.

Taproots.tapout — Method

tapout(sink::Function, taproot::Taproot)

This allows you to transform a Taproots.Taproot struct back into one of your own structs. The sink function needs to be able to transform a single node of a Taproot into a node in your struct. The sink follows the following structure: sink(data, children) which returns whatever type you want.

As a simple example, let's say we have

struct YourType printablename::String id::UUID data children::Vector{YourType} end Taproots.children(x::YourType) = x.children Taproots.data(x::YourType) = (x.printablename, x.id, x.data)

Then you can convert that to a Taproots.taproot as follows:

yourtype = ... yourtaproot = tapin(yourtype) yourmodifiedtype = tapout((data, children) -> YourType(data.printablename, data.id, data.data, children), modified_taproot)

In case you have multiple types in your struct chain, you can say:

function sink(data, children) if data isa String return data else return YourType(data.printable_name, data.id, data.data, children) end end

Taproots.topdown — Method

topdown(x)

This creates a lazy iterator as a topdown (level-order) breadth-first search of your custom Taproot. The top parents are iterated first, and then the next level down, and so on. All parents are guaranteed to be iterated on before the leaves.

Usage

for x in topdown(x) print(x) end

collect(topdown(x)) <: Vector

Taproots.tracepairs — Method

tracepairs(x)

This creates a lazy iterator for all the traces and nodes of a taproot. This will always be in the preorder (depth-first search).

Taproots.traces — Method

traces(x)

This creates a lazy iterator for all the traces (vectors of indices) needed to get from the root to one of the children. This will always be in the preorder (depth-first search).

Taproots.@sprout — Macro

@sprout YourType

This just sets AbstractTrees.children(x::YourType) = Taproots.children(x::YourType). Must be used before you can bloom or @bloom.